Put dataclasses in separate file
